Split CST builder into modules
